+++ doc/changelog.rst
@@ -6,8 +6,11 @@ What's New
 * Fixed NameError with PhantomJS proxy on Python 3.2.
 * Fixed PhantomJS stop waiting for page load too early.
 * Fixed ``--page-requisites`` exceeding ``--level``.
+* Fixed ``--no-verbose`` not providing informative messages and behaving like ``--quiet``.
 * Fixed infinite page requisite recursion when using ``--span-hosts-allow page-requisites``.
 * Added ``--page-requisites-level``. The default max recursion depth on page requisites is now 5.
+* Added ``--very-quiet``.
+* ``--no-verbose`` is defaulted when ``--concurrent`` is 2 or greater.
 
 Database Schema: